pyproject.toml
[tool.poetry]
name = "seerlord-ai"
version = "0.1.0"
description = "Modular AI Platform using Micro-Kernel + Plugin architecture"
authors = ["Admin <admin@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "*"
langchain-openai = "*"
fastapi = "*"
uvicorn = "*"
pydantic-settings = "*"
asyncpg = "*"
psycopg2-binary = "*"
langserve = "*"
langgraph-checkpoint-postgres = "^3.0.2"
psycopg-pool = "^3.3.0"
psycopg-binary = "^3.3.2"
sse-starlette = "^3.0.3"
loguru = "^0.7.2"
markitdown = {version = ">=0.0.1a3", extras = ["docx"]}
python-docx = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
